% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_to_binary.R
\name{check_to_binary}
\alias{check_to_binary}
\title{Convert REDCap Select all that apply "label" to "binary" intended to modify
input for tbl_summary() calls.}
\usage{
check_to_binary(df, cols = everything(), val_to_1 = "Checked")
}
\arguments{
\item{df}{A data frame containing character columns of select all that apply}

\item{val_to_1}{A character vector specifying the character value to convert to 1}
}
\value{
A data frame with average scores
}
\description{
Convert REDCap Select all that apply "label" to "binary" intended to modify
input for tbl_summary() calls.
}
\examples{
data \%>\%
 select(record_id, expect_certain_number:appreciation) \%>\%
 score_ic_scale(., "record_id")
}
