% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gt_tbl.R
\name{format_gt_tbl}
\alias{format_gt_tbl}
\title{format_gt_tbl()}
\usage{
format_gt_tbl(df)
}
\arguments{
\item{df}{a dataframe or tibble of statistical output such as that from
broom::tidy()}

\item{tol}{a tolerance level}
}
\value{
a gt() table
}
\description{
Format a dataframe or tibble of statistical output for dispaly in a rendered
Quarto document. More specifically this function rounds digits and formats
p.values to < 0.001 if value is met.
}
