% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fx_sz_tbl.R
\name{get_fx_sz_tbl}
\alias{get_fx_sz_tbl}
\title{get_fx_sz_tbl}
\usage{
get_fx_sz_tbl(alphas, sample_sizes, ratios, std, test_type)
}
\arguments{
\item{alphas}{A numeric vector of alpha levels to}

\item{sample_sizes}{A vector of numeric sample sizes to determine the
detectable effect sizes for}

\item{ratios}{A vector of sample size ratios to account for
attrition/unbalanced groups}

\item{std}{A numeric value representing the estimated standard deviation}

\item{test_type}{A string of either "paired" or "two.sample"}
}
\value{
A data frame of detectable effect sizes given power = 80\%.
}
\description{
Generate a table of detectable effect sizes given a sample size and standard
deviation
}
