% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_redcap_report.R
\name{pull_redcap_report}
\alias{pull_redcap_report}
\title{pull_redcap_report}
\usage{
pull_redcap_report(token, report_id, value_type, header_type)
}
\arguments{
\item{token}{A RedCap API token as a string corresponding to the RedCap
project of interest. The token can be saved as an environmental variable
in Windows and retrieved via Sys.getenv("\link{environmental variable name}") to
prevent uploading the token to GitHub.}

\item{report_id}{A numeric report id as a string that corresponds to a given
report in the RedCap project of interest}

\item{value_type}{Options are "label" or "raw". "label" returns the textual
version of the RedCap field responses such as "5 - Strongly agree" or
"Male" for fixed responses and "Checked"/"Unchecked" for select all that
apply. "raw" returns numeric values such as 5,}

\item{header_type}{Options are "label" or "raw". "raw" pulls the field names
as found in the RedCap project, while "label" pulls the questions that
correspond to the field eg. "Rate on a scale of 1-10 your level of pain."}
}
\value{
a dataframe
}
\description{
Pulls a report from a RedCap project given a project token, report id, and a
value type.
}
\examples{
# Basic example usage
data <- pull_redcap_report(Sys.getenv("LC_patient"), "81729739", "label")
 where "LC_patient" is a Windows environmental variable containing the 
 RedCap project API.

}
