% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_implementation_scale.R
\name{score_ic_scale}
\alias{score_ic_scale}
\title{Score Implementation Climate Scale}
\usage{
score_ic_scale(df, id)
}
\arguments{
\item{df}{A data frame containing 6 numeric columns and an identifier column.}

\item{id}{A character vector specifying the name of the identifier column.}
}
\value{
A data frame with average scores
}
\description{
Score Implementation Climate Scale
}
\examples{
data \%>\%
 select(record_id, expect_certain_number:appreciation) \%>\%
 score_ic_scale(., "record_id")
}
