% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_practice_culture_scale.R
\name{score_pc_scale}
\alias{score_pc_scale}
\title{Score Practice Culture Scale}
\usage{
score_pc_scale(df, id)
}
\arguments{
\item{df}{A data frame containing 22 numeric columns and an identifier column.
The 22 columns must be in order of how the scale was intended to be
administered}

\item{id}{A character vector specifying the name of the identifier column.}
}
\value{
A data frame with subscale scores and overall culture score.
}
\description{
This function scores the 22-item practice culture scale from survey data. Input
data frame must be in the following order:
\enumerate{
\item After making a change, we discuss what worked and what didn't.
\item My opinion is valued by others in this practice.
\item People in this practice understand how their jobs fit into the rest of this practice.
\item This practice is almost always in chaos.
\item I can rely on the other people in this practice to do their jobs well.
\item This practice puts a great deal of effort into improving the quality of care.
\item This practice encourages everybody's input for making changes.
\item We regularly take time to consider ways to improve how we do things.
\item The practice leadership makes sure that we have the time and space necessary to discuss changes to improve care.
\item This practice is very disorganized.
\item When there is conflict or tension in this practice, those involved are encouraged to talk about it.
\item People in this practice are thoughtful about how they do their jobs.
\item This practice uses data and information to improve the work of the practice.
\item Our practice encourages people to share their ideas about how to improve things.
\item People in this practice pay attention to how their actions affect others in the practice.
\item The leadership in this practice is available to discuss work related problems.
\item When we experience a problem in the practice we make a serious effort to figure out what's really going on.
\item Our practice has recently been very stable.
\item Things have been changing so fast in our practice that it is hard to keep up with what is going on.
\item The leadership of this practice is good at helping us to make sense of problems or difficult situations.
\item Most of the people who work in our practicee seem to enjoy their work.
\item The practice leadership promotes an environment that is an enjoyable place to work.
}
}
\details{
The practice culture is composed of the following three sub-scales that are
scored as follows.
\enumerate{
\item Improvement and Change Culture
\itemize{
\item 1, 6:9, 13:14, 16:17, 20
\item Scoring: 25*((q1 + q6 + q7 + q8 + q9 + q13 + q14 +q16 + q17 + q20) -10)/10
}
\item Work Relationships
\itemize{
\item 2:3, 5, 11:12, 15, 21:22
\item Scoring: 25*((q2 + q3 + q5 + q11 + q12 + q15 + q21 + q22) – 8) /8
}
\item Chaos
\itemize{
\item 4, 10, 18:19
\item Scoring:  25*((q4 + q10 + (6-q18) + q19) - 4)/4
\item N.b.: Item 18 (stability) is reverse scored
}
}
}
\examples{
data \%>\%
 select(record_id, after_change:environment) \%>\%
 score_pc_scale(., "record_id")
}
