% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable.R
\name{summarize_variable}
\alias{summarize_variable}
\title{summarize_variable()}
\usage{
summarize_variable(df, var_name)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{column}{a variable name in the input data frame, no quotes needed}
}
\value{
a tibble
}
\description{
Summarize a numeric variable. Calculates the non-missing, min, max, median,
mean, sd, and se of a numeric variable.
}
\examples{
# Basic example usage
summarize_variable(heights_df, height)

}
